Add files via upload
Table of Contents:
Dataset Preparation: The first step is the Dataset Preparation step which includes the process of loading a dataset and performing basic pre-processing. The dataset is then splitted into train and validation sets.
Feature Engineering: The next step is the Feature Engineering in which the raw dataset is transformed into flat features which can be used in a machine learning model. This step also includes the process of creating new features from the existing data.
Model Training: The final step is the Model Building step in which a machine learning model is trained on a labelled dataset.
Improve Performance of Text Classifier: In this article, we will also look at the different ways to improve the performance of text classifiers.